refactor: icons + imports + delete components
