--- src/components/ui/DatePicker/DatePicker.tsx
@@ -1,3 +1,0 @@
-export const DatePicker = () => {
-  return <></>
-}
